Map updated and logic ready for welcome message
Inside of the BeginFormController there is a new method used to display the welcome message. The logic is there, commented out, but has not been implemented yet. Also, there have been extensive changes to the MapController and the map that we are using now, MapTest.hmtl. The name has no been refactored but that''s the one it uses. Also, there is a text file being used to share visitor data, if you use it try not to wipe the data from it before it reaches where it needs to go in BeginFormController. This is still a work in progress and the map has not been finished, it merely has its functionality *mostly* complete. Thanks guys.